+++ GetWired New Firmware/InternalTemp.cpp
