+++ GetWired New Firmware/PowerSensor.cpp
