+++ GetWired New Firmware/RShutterControl.cpp
